Fix bug resizing const Mat in detect_digits.cpp; refactor to separate sudokusolver package and gosudoku package with main entry point
